fix: gate parallel code behind its feature
